fix(fixture): The properties are loaded from the store
In order to use the fixture-detail-view to edit it's properties it was necesarray to also save the
properties into the store. This makes it possible to also rerender to the last values when editing
another fixture. Also it's now possible to edit all properties and meta data of the fixture.

Fixes #80, #8, #74