fix(dependency): Updated to lit-html 1.1.2 & lit-element 2.1.0
In order to run luminave with the latest version of lit-html & lit-element, we had to get rid of
some deprecated code. classMap.js was renamed into class-map.js. when.js was removed. The directives
are working differently now, so we had to update defaultValue(). Also updated the README to make
clear that we recommend to use npm to install dependencies.

fix #114